Add breakpoint distance method to comparison validation tools.
Plots the combined distance to the breakpoints left and right of each intersection segment (after merging the two profiles) and assigns to one of four categories based on size and placement of original segments (relative to merged segment). Had to update get_differences_from_intervals with optional argument to access the data for this.